interview-service-ai/InterviewQuestionService/requirements.txt
fastapi==0.123.0
uvicorn==0.36.0
mongoengine==0.29.1
pymongo==4.15.1
python-dotenv==1.1.1
requests==2.32.5
kafka-python==2.2.15
python-docx==1.2.0
PyPDF2==3.0.1
moviepy==1.0.3
google-generativeai==0.8.5
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
pydantic==2.11.7